Refactor: Major refactor to the SignInPage's onChange functions. Was a test to see how to handle the UserProfile page. Works, will continue refactor.
